Extract input validation to a dedicated subdirectory
Extract input validation to a dedicated subdirectory
